ms_write_file/requirements.txt
fastapi
uvicorn
python-dotenv
celery
redis